+++ src/components/EditableCellRelational.vue
@@ -36,6 +36,7 @@
         :item="item"
         :field="actualFieldKey"
         :edit-mode="props.editMode"
+        :alignment="props.align"
       />
       <!-- Use custom ImageCell for image fields -->
       <ImageCell
@@ -51,6 +52,7 @@
         :value="value"
         :item="item"
         :field="actualFieldKey"
+        :alignment="align"
       />
       <!-- Use custom RelationalCell for relational fields -->
       <RelationalCell